refactor(direct-control): extract health result type
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves Civ7DirectControlHealth into src/session/types.ts while keeping public facade type re-exports in index.ts.

Preserves the health result shape and does not change endpoint discovery, state selection, health polling, error classification, session/socket behavior, or runtime-status projection. Health source/session orchestration, public procedure schemas, operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- session.test.ts public-api.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check
- ownership scan for health result type definition

Local package/type relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
